Methods

Problem

Missing docstring.

Missing docstring for Problem. Check Documenter's build log for details.

TORA.fix_joint_positions! — Function

fix_joint_positions!(problem, robot, knot, q)

Returns a.

TORA.fix_joint_velocities! — Function

fix_joint_velocities!(problem, robot, knot, v)

To do.

TORA.fix_joint_torques! — Function

fix_joint_torques!(problem, robot, knot, τ)

To do.

TORA.constrain_ee_position! — Function

constrain_ee_position!(problem, knot, position)

To do.

Missing docstring.

Missing docstring for show_problem_info. Check Documenter's build log for details.

Solver Log

TORA.SolverLog — Type

SolverLog(N::Int)

Creates an empty NLP solver log.

Arguments

N: the total number of decision variables of the NLP.

TORA.update! — Function

update!(log::SolverLog,
        x::Array{Float64,2},
        abs_feas_error::Array{Float64,1},
        abs_opt_error::Array{Float64,1},
        obj_value::Array{Float64,1},
        fc_evals::Array{Float64,1},
        ga_evals::Array{Float64,1})

Updates the log.

Arguments

log: instance of a SolverLog.

x: an intermediate iterate point.

abs_feas_error: absfeaserror at the current point.

abs_opt_error: absopterror at the current point.

obj_value: obj_value at the current point.

fc_evals: the number of function evaluations requested by the NLP solver so far.

ga_evals: the number of gradient evaluations requested by the NLP solver so far.

Missing docstring.

Missing docstring for length. Check Documenter's build log for details.

Missing docstring.

Missing docstring for save. Check Documenter's build log for details.

Missing docstring.

Missing docstring for load!. Check Documenter's build log for details.

Plot Utilities

Missing docstring.

Missing docstring for plot_results. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plot_log. Check Documenter's build log for details.

Index

TORA.SolverLog

TORA.constrain_ee_position!

TORA.fix_joint_positions!

TORA.fix_joint_torques!

TORA.fix_joint_velocities!

TORA.update!