Robot

TORA.Robot — Type

A Robot represents a mechanism, its state, and other relevant objects.

Relevant fields in Robot

The arrays q_lo, q_hi, v_lo, v_hi, τ_lo, τ_hi store the lower and upper bounds of the joint positions, velocities, and torques, respectively.

The number of generalized coordinates, generalized velocities, and actuated joints are stored in n_q, n_v, n_τ, respectively.

The end-effector frame is stored in frame_ee.

TORA.Robot — Method

Robot(urdfpath, mechanism, frame_ee, mvis)

Create a new Robot.

Index

TORA.Robot

TORA.Robot

TORA.create_robot_kuka_iiwa_14

Methods

TORA.create_robot_kuka_iiwa_14 — Method

create_robot_kuka_iiwa_14(vis)

Create a new KUKA LBR iiwa 14 robot.