Solver Log

TORA.SolverLog — Type

During optimization, a SolverLog is used to store convergence information.

A SolverLog can log the following:

x::Matrix{Float64}: the history of the decision variables for every iteration.

abs_feas_error::Vector{Float64}: the absolute feasibility error evolution.

abs_opt_error::Vector{Float64}: the absolute optimality error evolution.

obj_value::Vector{Float64}: the evolution of the objective function value.

fc_evals::Vector{Float64}: the evolution of the number of function evaluations.

ga_evals::Vector{Float64}: the evolution of the number of gradient evaluations.

nStatus::Int: the solution status return code.

The actual values stored in a SolverLog depend on the implementation of the callback used by each solver for every new point. Using Ipopt, the callback is set with Ipopt.setIntermediateCallback. Using Knitro, the callback is set with KNITRO.KN_set_newpt_callback.

TORA.SolverLog — Method

SolverLog(N)

Create an empty NLP solver log.

Arguments

N::Int: the total number of decision variables of the NLP.

See also: update!

Index

TORA.SolverLog

TORA.SolverLog

Base.length

TORA.load!

TORA.save

TORA.update!

Methods

Base.length — Method

length(log) -> Integer

Return the length of a SolverLog.

The length of a SolverLog is the same as the number of iterations taken to solve the problem.

TORA.load! — Method

load!(log, file)

Load an .npz file and store it in a SolverLog object.

See also: save

TORA.save — Method

save(log, file)

Save a SolverLog to an .npz file.

See also: load!

TORA.update! — Method

update!(log; <keyword arguments>)

Update a SolverLog object.

Keyword arguments

x=zeros(size(log.x, 1)): an intermediate iterate point.

abs_feas_error=0: the absolute feasibility error at the current point.

abs_opt_error=0: the absolute optimality error at the current point.

obj_value=0: the value of the objective function at the current point.

fc_evals=0: the number of function evaluations requested by the NLP solver so far.

ga_evals=0: the number of gradient evaluations requested by the NLP solver so far.