Problem

TORA.Problem — Type

A Problem stores the high-level description of a motion planning problem.

This is the main structure of the package. The idea is for users to start with an empty problem for a given robot. Then, in order to specify the requirements for the motion they want to compute, they can add different types of constraints, one by one. See Creating a Problem and Defining Constraints for detailed examples.

Eventually, the problem becomes a complex description with all the requirements. Once the description is finalized, it can be optimized. This happens in one of the TORA.solve_with_* functions, which does the heavy-lifting of converting a Problem (the high-level description) into an NLP problem that can be solved using off-the-shelf solvers, such as Ipopt and Knitro.

See also: solve_with_ipopt, solve_with_knitro

TORA.Problem — Method

Problem(robot, num_knots, dt)

Create an empty Problem.

Arguments

robot::Robot: the robot associated with this Problem.

num_knots::Int: the number of knots used to represent the motion.

dt::Float64: the time step duration between each two consecutive knots.

See also: show_problem_info

Index

TORA.Problem

TORA.Problem

TORA.constrain_ee_position!

TORA.export_trajectory

TORA.fix_joint_positions!

TORA.fix_joint_torques!

TORA.fix_joint_velocities!

TORA.show_problem_info

Methods

TORA.constrain_ee_position! — Method

constrain_ee_position!(problem, knot, position)

Fix the end-efector positions of the robot at a specific knot.

See also: fix_joint_positions!, fix_joint_velocities!, fix_joint_torques!

TORA.export_trajectory — Method

export_trajectory(file, problem, robot, x)

Export a trajectory to a .npz file.

TORA.fix_joint_positions! — Method

fix_joint_positions!(problem, robot, knot, q)

Fix the joint positions of the robot at a specific knot.

See also: fix_joint_velocities!, fix_joint_torques!

TORA.fix_joint_torques! — Method

fix_joint_torques!(problem, robot, knot, τ)

Fix the joint torques of the robot at a specific knot.

See also: fix_joint_positions!, fix_joint_velocities!

TORA.fix_joint_velocities! — Method

fix_joint_velocities!(problem, robot, knot, v)

Fix the joint velocities of the robot at a specific knot.

See also: fix_joint_positions!, fix_joint_torques!

TORA.show_problem_info — Method

show_problem_info(problem)

Output a summary of the problem, including the number of knots with constraints.