Robot

TORA.Robot — Type

A Robot represents a mechanism, its state, and other relevant objects.

Relevant fields in Robot

The arrays q_lo, q_hi, v_lo, v_hi, τ_lo, τ_hi store the lower and upper bounds of the joint positions, velocities, and torques, respectively.

The number of generalized coordinates, generalized velocities, and actuated joints are stored in n_q, n_v, n_τ, respectively.

The end-effector frame is stored in frame_ee.

TORA.Robot — Method

Robot(urdfpath, mechanism, frame_ee, mvis)

Create a new Robot.

Index

TORA.Robot

TORA.Robot

TORA.create_robot_franka

TORA.create_robot_kinova_gen2

TORA.create_robot_kinova_gen3

TORA.create_robot_kuka

TORA.create_robot_ur

Methods

TORA.create_robot_franka — Method

create_robot_franka(model, vis)

Create a new Franka Emika robot.

TORA.create_robot_kinova_gen2 — Method

create_robot_kinova_gen2(model, vis)

Create a new Kinova Gen2 robot.

TORA.create_robot_kinova_gen3 — Method

create_robot_kinova_gen3(model, vis)

Create a new Kinova Gen3 lite robot.

TORA.create_robot_kuka — Method

create_robot_kuka(model, vis)

Create a new KUKA LBR iiwa 7/14 robot.

TORA.create_robot_ur — Method

create_robot_ur(ur_type, vis)

Create a new Universal Robots robot.